# Optimize haste map data structure for serialization/deserialization.