--- a/packages/jest-haste-map/src/__tests__/index.test.js
+++ b/packages/jest-haste-map/src/__tests__/index.test.js
@@ -340,17 +340,17 @@ describe('HasteMap', () => {
 
       expect(data.files).toEqual(
         createMap({
-          'fruits/Banana.js': ['Banana', 32, 42, 1, ['Strawberry'], null],
-          'fruits/Pear.js': ['Pear', 32, 42, 1, ['Banana', 'Strawberry'], null],
-          'fruits/Strawberry.js': ['Strawberry', 32, 42, 1, [], null],
-          'fruits/__mocks__/Pear.js': ['', 32, 42, 1, ['Melon'], null],
+          'fruits/Banana.js': ['Banana', 32, 42, 1, 'Strawberry', null],
+          'fruits/Pear.js': ['Pear', 32, 42, 1, 'Banana\tStrawberry', null],
+          'fruits/Strawberry.js': ['Strawberry', 32, 42, 1, '', null],
+          'fruits/__mocks__/Pear.js': ['', 32, 42, 1, 'Melon', null],
           // node modules
           'fruits/node_modules/fbjs/lib/flatMap.js': [
             'flatMap',
             32,
             42,
             1,
-            [],
+            '',
             null,
           ],
           'fruits/node_modules/react/React.js': [
@@ -358,10 +358,10 @@ describe('HasteMap', () => {
             32,
             42,
             1,
-            ['Component'],
+            'Component',
             null,
           ],
-          'vegetables/Melon.js': ['Melon', 32, 42, 1, [], null],
+          'vegetables/Melon.js': ['Melon', 32, 42, 1, '', null],
         }),
       );
 
@@ -410,18 +410,11 @@ describe('HasteMap', () => {
 
           // The node crawler returns "null" for the SHA-1.
           data.files = createMap({
-            'fruits/Banana.js': ['Banana', 32, 42, 0, ['Strawberry'], null],
-            'fruits/Pear.js': [
-              'Pear',
-              32,
-              42,
-              0,
-              ['Banana', 'Strawberry'],
-              null,
-            ],
-            'fruits/Strawberry.js': ['Strawberry', 32, 42, 0, [], null],
-            'fruits/__mocks__/Pear.js': ['', 32, 42, 0, ['Melon'], null],
-            'vegetables/Melon.js': ['Melon', 32, 42, 0, [], null],
+            'fruits/Banana.js': ['Banana', 32, 42, 0, 'Strawberry', null],
+            'fruits/Pear.js': ['Pear', 32, 42, 0, 'Banana/Strawberry', null],
+            'fruits/Strawberry.js': ['Strawberry', 32, 42, 0, '', null],
+            'fruits/__mocks__/Pear.js': ['', 32, 42, 0, 'Melon', null],
+            'vegetables/Melon.js': ['Melon', 32, 42, 0, '', null],
           });
 
           return Promise.resolve({
@@ -446,7 +439,7 @@ describe('HasteMap', () => {
               32,
               42,
               1,
-              ['Strawberry'],
+              'Strawberry',
               '7772b628e422e8cf59c526be4bb9f44c0898e3d1',
             ],
             'fruits/Pear.js': [
@@ -454,7 +447,7 @@ describe('HasteMap', () => {
               32,
               42,
               1,
-              ['Banana', 'Strawberry'],
+              'Banana\tStrawberry',
               '89d0c2cc11dcc5e1df50b8af04ab1b597acfba2f',
             ],
             'fruits/Strawberry.js': [
@@ -462,7 +455,7 @@ describe('HasteMap', () => {
               32,
               42,
               1,
-              [],
+              '',
               'e8aa38e232b3795f062f1d777731d9240c0f8c25',
             ],
             'fruits/__mocks__/Pear.js': [
@@ -470,7 +463,7 @@ describe('HasteMap', () => {
               32,
               42,
               1,
-              ['Melon'],
+              'Melon',
               '8d40afbb6e2dc78e1ba383b6d02cafad35cceef2',
             ],
             'vegetables/Melon.js': [
@@ -478,7 +471,7 @@ describe('HasteMap', () => {
               32,
               42,
               1,
-              [],
+              '',
               'f16ccf6f2334ceff2ddb47628a2c5f2d748198ca',
             ],
           }),
@@ -626,7 +619,7 @@ describe('HasteMap', () => {
               32,
               42,
               1,
-              ['Blackberry'],
+              'Blackberry',
               null,
             ],
             'fruits/Strawberry.ios.js': [
@@ -634,10 +627,10 @@ describe('HasteMap', () => {
               32,
               42,
               1,
-              ['Raspberry'],
+              'Raspberry',
               null,
             ],
-            'fruits/Strawberry.js': ['Strawberry', 32, 42, 1, ['Banana'], null],
+            'fruits/Strawberry.js': ['Strawberry', 32, 42, 1, 'Banana', null],
           }),
         );
 
@@ -724,14 +717,7 @@ describe('HasteMap', () => {
             expect(useBuitinsInContext(data.clocks)).toEqual(mockClocks);
 
             const files = new Map(initialData.files);
-            files.set('fruits/Banana.js', [
-              'Banana',
-              32,
-              42,
-              1,
-              ['Kiwi'],
-              null,
-            ]);
+            files.set('fruits/Banana.js', ['Banana', 32, 42, 1, 'Kiwi', null]);
 
             expect(useBuitinsInContext(data.files)).toEqual(files);
 
@@ -1102,7 +1088,7 @@ describe('HasteMap', () => {
 
         expect(data.files).toEqual(
           createMap({
-            'fruits/Banana.js': ['Banana', 32, 42, 1, ['Strawberry'], null],
+            'fruits/Banana.js': ['Banana', 32, 42, 1, 'Strawberry', null],
           }),
         );
 
@@ -1136,7 +1122,7 @@ describe('HasteMap', () => {
 
         expect(data.files).toEqual(
           createMap({
-            'fruits/Banana.js': ['Banana', 32, 42, 1, ['Strawberry'], null],
+            'fruits/Banana.js': ['Banana', 32, 42, 1, 'Strawberry', null],
           }),
         );
       });