--- a/packages/jest-haste-map/src/constants.ts
+++ b/packages/jest-haste-map/src/constants.ts
@@ -18,6 +18,9 @@
 import {HType} from './types';
 
 const constants: HType = {
+  /* dependency serialization */
+  DEPENDENCY_DELIM: ':',
+
   /* file map attributes */
   ID: 0,
   MTIME: 1,