--- a/packages/jest-haste-map/src/constants.ts
+++ b/packages/jest-haste-map/src/constants.ts
@@ -19,7 +19,7 @@ import {HType} from './types';
 
 const constants: HType = {
   /* dependency serialization */
-  DEPENDENCY_DELIM: ':',
+  DEPENDENCY_DELIM: '\0',
 
   /* file map attributes */
   ID: 0,