--- a/packages/jest-haste-map/src/crawlers/__tests__/node.test.js
+++ b/packages/jest-haste-map/src/crawlers/__tests__/node.test.js
@@ -133,9 +133,9 @@ describe('node crawler', () => {
 
       expect(hasteMap.files).toEqual(
         createMap({
-          'fruits/strawberry.js': ['', 32, 42, 0, [], null],
-          'fruits/tomato.js': ['', 33, 42, 0, [], null],
-          'vegetables/melon.json': ['', 34, 42, 0, [], null],
+          'fruits/strawberry.js': ['', 32, 42, 0, '', null],
+          'fruits/tomato.js': ['', 33, 42, 0, '', null],
+          'vegetables/melon.json': ['', 34, 42, 0, '', null],
         }),
       );
 
@@ -151,9 +151,9 @@ describe('node crawler', () => {
     nodeCrawl = require('../node');
 
     // In this test sample, strawberry is changed and tomato is unchanged
-    const tomato = ['', 33, 42, 1, [], null];
+    const tomato = ['', 33, 42, 1, '', null];
     const files = createMap({
-      'fruits/strawberry.js': ['', 30, 40, 1, [], null],
+      'fruits/strawberry.js': ['', 30, 40, 1, '', null],
       'fruits/tomato.js': tomato,
     });
 
@@ -166,7 +166,7 @@ describe('node crawler', () => {
     }).then(({hasteMap, removedFiles}) => {
       expect(hasteMap.files).toEqual(
         createMap({
-          'fruits/strawberry.js': ['', 32, 42, 0, [], null],
+          'fruits/strawberry.js': ['', 32, 42, 0, '', null],
           'fruits/tomato.js': tomato,
         }),
       );
@@ -186,9 +186,9 @@ describe('node crawler', () => {
     // In this test sample, previouslyExisted was present before and will not be
     // when crawling this directory.
     const files = createMap({
-      'fruits/previouslyExisted.js': ['', 30, 40, 1, [], null],
-      'fruits/strawberry.js': ['', 33, 42, 0, [], null],
-      'fruits/tomato.js': ['', 32, 42, 0, [], null],
+      'fruits/previouslyExisted.js': ['', 30, 40, 1, '', null],
+      'fruits/strawberry.js': ['', 33, 42, 0, '', null],
+      'fruits/tomato.js': ['', 32, 42, 0, '', null],
     });
 
     return nodeCrawl({
@@ -200,13 +200,13 @@ describe('node crawler', () => {
     }).then(({hasteMap, removedFiles}) => {
       expect(hasteMap.files).toEqual(
         createMap({
-          'fruits/strawberry.js': ['', 32, 42, 0, [], null],
-          'fruits/tomato.js': ['', 33, 42, 0, [], null],
+          'fruits/strawberry.js': ['', 32, 42, 0, '', null],
+          'fruits/tomato.js': ['', 33, 42, 0, '', null],
         }),
       );
       expect(removedFiles).toEqual(
         createMap({
-          'fruits/previouslyExisted.js': ['', 30, 40, 1, [], null],
+          'fruits/previouslyExisted.js': ['', 30, 40, 1, '', null],
         }),
       );
     });
@@ -228,8 +228,8 @@ describe('node crawler', () => {
     }).then(({hasteMap, removedFiles}) => {
       expect(hasteMap.files).toEqual(
         createMap({
-          'fruits/directory/strawberry.js': ['', 33, 42, 0, [], null],
-          'fruits/tomato.js': ['', 32, 42, 0, [], null],
+          'fruits/directory/strawberry.js': ['', 33, 42, 0, '', null],
+          'fruits/tomato.js': ['', 32, 42, 0, '', null],
         }),
       );
       expect(removedFiles).toEqual(new Map());
@@ -252,8 +252,8 @@ describe('node crawler', () => {
     }).then(({hasteMap, removedFiles}) => {
       expect(hasteMap.files).toEqual(
         createMap({
-          'fruits/directory/strawberry.js': ['', 33, 42, 0, [], null],
-          'fruits/tomato.js': ['', 32, 42, 0, [], null],
+          'fruits/directory/strawberry.js': ['', 33, 42, 0, '', null],
+          'fruits/tomato.js': ['', 32, 42, 0, '', null],
         }),
       );
       expect(removedFiles).toEqual(new Map());