--- a/packages/jest-haste-map/src/crawlers/watchman.ts
+++ b/packages/jest-haste-map/src/crawlers/watchman.ts
@@ -232,7 +232,7 @@ export = async function watchmanCrawl(
           ];
         } else {
           // See ../constants.ts
-          nextData = ['', mtime, size, 0, [], sha1hex];
+          nextData = ['', mtime, size, 0, '', sha1hex];
         }
 
         const mappings = options.mapper ? options.mapper(filePath) : null;