## Summary

This simple PR is the result of a ridiculous number of attempts to specifically optimize the serialization and deserialization of the haste map followed by a simple realization: creating arrays is expensive. :)

Currently, we store each file's dependencies as an expensive array. However, we don't typically access many file dependencies, so it's not time well-spent.

I've modified the data structure to store the file dependencies as a much cheaper string (tab separated) that is deserialized to an array on-demand. No change to the public interface, but now we're not doing any unnecessary work.

I profiled both the full Jest run time and the read/persist time to ensure I had a clear view into the characteristics of this change. It's somewhere in the neighborhood of a 10%~ startup time improvement at FB, which is significant.

Benchmarking against a generated benchmark hash map of 300,000 files:

before--
read: 2,291ms
persist: 2,413ms
total: 4,704ms

after--
read: 1,852ms
persist: 1,539ms
total: 3,391ms

delta: 1,313ms

## Test plan

- Updated tests
- All tests pass
- Manually benchmarked