@jeysal 2019-03-20
These numbers sound great, thanks! I think we should use the colon (`:`) as the delimiter though. Afaik only `NUL` and `:` are forbidden in UNIX paths, a tab could occur in a path normally.