@jeysal 2019-03-20
Try e.g.
```js
require('fs').writeFileSync('/tmp/\t', 'asdf')
```