@SimenB 2019-03-20
We talked about it, but just so it's written down a `:` might not work on windows if the file is on another drive (`D:/blabla`)