@rubennorte 2019-03-20
> These numbers sound great, thanks! I think we should use the colon (:) as the delimiter though. Afaik only NUL and : are forbidden in UNIX paths, a tab could occur in a path normally.

Yeah, I'd recommend using `\0`.

@scotthovestadt did you measure the impact in building the haste map or the full startup time? I imagine this can have an impact in the time to determine the tests to run (see https://github.com/facebook/jest/blob/master/packages/jest-resolve-dependencies/src/index.ts) as it resolves the dependencies of all the modules in the haste map. Maybe it'd be worth it to memoize the array (even in the same object, as it won't be persisted again).