@scotthovestadt 2019-03-20
@jeysal Will extract the delim to a var and use `:`.

Considered memoizing but don't think it'll be worth it with the current access patterns.

I did test the haste map generation and while there might be a small hit it wasn't actually measurable because:
1) It's a relatively variable time already (it has a range of a couple seconds once you hit the threshold I'm benchmarking at).
2) The persist at the end is faster so that might be making up for the slower generation.

In any case, it can be optimized away by doing the join in the worker thread (which would introduce the complexity of detecting which type to avoid a breaking change). I didn't go there yet because I wasn't able to measure a difference by doing that... only pushing up changes where I can measure a significant difference. :)