@rubennorte 2019-03-20
@scotthovestadt my point is that if you only measure haste map build time you won't see that the cost might be transferring to a later stage in the process (in this case test determination). I'd try to measure the time it takes to get the list of tests that need to run when you use `--findRelatedTests` (see https://github.com/facebook/jest/blob/master/packages/jest-core/src/SearchSource.ts#L158). That should give you a good measure about the impact this could have in finding inverse dependencies.